refactor(wincon): Pull out platform-specific code
